Updated the Existing http request for adding firms
Removed http request from FirmController class, and put it in the admin
class. Re-named the /admins post mapping to /admins-create-admin. Renamed
the /firms post mapping to /admins-create-firm. Created a new FirmInput
class. Added a requirement that a valid admin's username and password
must be sent in the post call's request body.